AWS_REGION: MY_AWS_REGION
name: Configure AWS credentials
USER AGREEMENT
Last update: 22th of April, 2025
Welcome to Wallet in Telegram. This User Agreement (the “Agreement”) governs
your use of Wallet in Telegram, a mini app that facilitates the exchange of
Supported Virtual Assets and Traded Assets; enables deposits, withdrawals,
transfers, and storage of Supported Virtual Assets, and provides access to a
peer-to-peer matching service (the “Service” or “Services”). 
1. Introduction
1.1. The Services are provided by Wallet Service Provider, (referred to as
“Wallet,” “us,” or “we”) to you, the “User” or “you”.
1.2. Wallet provides its Services independently and is not a subsidiary or an
affiliate of Telegram FZ-LLC.
1.3. By accessing and using Wallet, you agree to be bound by this Agreement,
along with any additional policies, including the Privacy Policy and terms
referenced herein, as well as the terms of any campaigns or additional
Services that Wallet may make available to you from time to time, even if not
explicitly referenced here. If you disagree with any part of this Agreement or
if the terms conflict with the laws applicable in your country of residence or
citizenship, you should not use Wallet and must cease using the Services.
1.4. We may update this Agreement from time to time. The revised Agreement
will take effect when posted in Wallet, unless otherwise specified. Continued
use of the Services after such updates will be considered acceptance of the
revised Agreement.
2. Risk Disclosure
2.1. The Services are currently not subject to oversight by any regulatory
authority. When using the Services, you will not have access to protections
typically available to consumers of regulated financial or similar services,
such as the right to file complaints with a financial ombudsman or similar
channels, or to rely on any statutory compensation schemes. You are solely
responsible for assessing the suitability and risks associated with the
Services.
2.2.Transactions involving virtual assets are inherently risky and can result in
significant financial loss. You should only invest amounts you can afford to
lose and seek independent financial and legal advice as needed. Ensure that
you understand the risks involved and verify that you are permitted to use
Wallet based on your legal jurisdiction.
2.3.Wallet disclaims any liability for significant price volatility, devaluation,
delisting, forks, airdrops, or other risks related to Virtual Assets transactions.